+++ assets/css/app.scss
@@ -161,8 +161,8 @@ form.link {
 
 @include media-breakpoint-down(sm) {
   #carousel-controls, .image-box img {
-    width: 200px;
-    height: 150px;
+    width: 240px;
+    height: 180px;
   }
 
   #carousel-controls .carousel-control-prev {
