+++ lib/melodica_inventory_web/controllers/admin/item_controller.ex
@@ -58,30 +58,6 @@ defmodule MelodicaInventoryWeb.Admin.ItemController do
   defp update_item(item, item_params) do
     Multi.new()
     |> Multi.update(:item, Item.changeset(item, item_params))
-    |> Multi.run(:image, fn state -> upload_images(state, item_params["image"]) end)
-  end
-
-  defp upload_images(%{item: %Item{}}, nil), do: {:ok, true}
-
-  defp upload_images(%{item: %Item{id: item_id}}, uploaded_files) do
-    Enum.map(uploaded_files, &upload_to_cloudex_and_save(&1, item_id))
-    |> format_response
-  end
-
-  defp upload_to_cloudex_and_save(%Plug.Upload{path: filename}, item_id) do
-    case Cloudex.upload(filename) do
-      [error: error] ->
-         :error
-      [ok: %Cloudex.UploadedImage{public_id: public_id}] ->
-        {:ok, Repo.insert(%Image{public_id: public_id, item_id: item_id})}
-    end
-  end
-
-  defp format_response(upload_images) do
-    if :error in upload_images do
-      {:error, "Cannot save image!"}
-    else
-      {:ok, true}
-    end
+    |> Multi.run(:image, fn state -> ImageOperations.upload_images(state, item_params["image"]) end)
   end
 end
