+++ lib/melodica_inventory_web/views/admin/item_view.ex
@@ -3,14 +3,5 @@ defmodule MelodicaInventoryWeb.Admin.ItemView do
 
   alias MelodicaInventory.Goods.Item
   alias Cloudex.Url
-
-  def cover_url(%Item{images: []}) do
-    []
-  end
-
-  def cover_url(%Item{images: images}) do
-    Enum.map(images, fn(image)  ->
-      {Url.for(image.public_id, %{width: 400, height: 300, crop: "limit"}), image.public_id}
-    end)
-  end
+  import MelodicaInventoryWeb.ItemView, only: [cover_urls: 1]
 end
