+++ lib/melodica_inventory_web/views/item_view.ex
@@ -5,17 +5,17 @@ defmodule MelodicaInventoryWeb.ItemView do
   import MelodicaInventoryWeb.ItemReservationView, only: [event_name: 1]
   import MelodicaInventoryWeb.EventView, only: [event_user: 1]
 
-  def cover_url(%Item{images: [], attachments: []}) do
+  def cover_urls(%Item{images: [], attachments: []}) do
     nil
   end
 
-  def cover_url(%Item{images: [], attachments: attachments}) do
+  def cover_urls(%Item{images: [], attachments: attachments}) do
     List.first(attachments).url
   end
 
-  def cover_url(%Item{images: images}) do
+  def cover_urls(%Item{images: images}) do
     Enum.map(images, fn(image)  ->
-      Url.for(image.public_id, %{width: 400, height: 300, crop: "limit"})
+      {Url.for(image.public_id, %{width: 400, height: 300, crop: "limit"}), image.public_id}
     end)
     |> Enum.with_index
   end
